Replace our stubby CNN with an outsourced Faster R-CNN
PyTorch was added as a backend with a small nested CNN model in
the past where a convolutional net would need a same-resolution
image and output a downsampled heat map of the object location.
Needless to say, this required a different model and training
procedures for each matched needle.

With the new approach we replace this with a state of the art
(bounding box regressor) model for object detection and associate
target patterns with COCO label classes. Further steps should help
out generalize this last part too but it is already really good
that the new model is completely outsourced and thus swappable.